+++ packages/client/src/rx/observable.test.ts
@@ -3,7 +3,8 @@ import { waitFor } from '@testing-library/dom';
 import { AnyRouter, TRPCError } from '../../../server/src';
 import { EventEmitter } from 'events';
 import { expectTypeOf } from 'expect-type';
-import { createChain, OperationLink, TRPCLink } from '../links/core';
+import { OperationLink, TRPCLink } from '../links/types';
+import { createChain } from '../links/internals/createChain';
 import { dedupeLink } from '../links/dedupeLink';
 import { splitLink } from '../links/splitLink';
 import { observable } from './observable';
